feat: add tab for activities filters
