+++ apps/web/lib/features/team/user-team-card/index.tsx
@@ -1,9 +1,11 @@
+'use client';
+
 import { secondsToTime } from '@app/helpers';
 import { useCollaborative, useTMCardTaskEdit, useTaskStatistics, useTeamMemberCard } from '@app/hooks';
 import { IClassName, IOrganizationTeamList } from '@app/interfaces';
 import { timerSecondsState } from '@app/stores';
 import { clsxm } from '@app/utils';
-import { Card, HorizontalSeparator, InputField, Text, Tooltip, VerticalSeparator } from 'lib/components';
+import { Card, HorizontalSeparator, InputField, Text, VerticalSeparator } from 'lib/components';
 import { DraggerIcon } from 'lib/components/svgs';
 import { TaskTimes, TodayWorkedTime } from 'lib/features';
 import { useTranslations } from 'next-intl';
@@ -14,35 +16,6 @@ import { UserInfo } from './user-info';
 import { UserTeamCardMenu } from './user-team-card-menu';
 import React from 'react';
 import UserTeamActivity from './user-team-card-activity';
-export function UserTeamCardHeader() {
-	const t = useTranslations();
-	return (
-		<div className="hidden sm:flex row font-normal justify-between pb-5 pt-8 hidde dark:text-[#7B8089]">
-			{/* <li className="pr-[50px]">{t('common.STATUS')}</li> */}
-			<div className="2xl:w-[20.625rem] text-center">{t('common.NAME')}</div>
-			<div className="w-1"></div>
-			<div className="2xl:w-80 3xl:w-[32rem] w-1/5 text-center">{t('common.TASK')}</div>
-			<div className="w-1"></div>
-			<Tooltip label={t('task.taskTableHead.TOTAL_WORKED_TODAY_HEADER_TOOLTIP')}>
-				<div className="2xl:w-48 3xl:w-[12rem] w-1/5 lg:px-4 px-2 flex flex-col justify-center text-center">
-					{t('task.taskTableHead.TASK_WORK.TITLE')}
-					<br />
-					{t('common.TASK')}
-				</div>
-			</Tooltip>
-			<div className="w-1"></div>
-			<div className=" text-center w-1/5 lg:px-3 2xl:w-52 3xl:w-64">{t('common.ESTIMATE')}</div>
-			<div className="w-1"></div>
-			<Tooltip label={t('task.taskTableHead.WORKED_ON_TASK_HEADER_TOOLTIP')}>
-				<div className="w-1/5 lg:px-3 2xl:w-52 3xl:w-64 text-center">
-					{t('task.taskTableHead.TOTAL_WORK.TITLE')}
-					<br />
-					{t('common.TODAY')}
-				</div>
-			</Tooltip>
-		</div>
-	);
-}
 
 type IUserTeamCard = {
 	active?: boolean;
@@ -63,10 +36,10 @@ export function UserTeamCard({
 	member,
 	publicTeam = false,
 	draggable = false,
-	onDragStart,
-	onDragEnd,
-	onDragEnter,
-	onDragOver,
+	onDragStart = () => null,
+	onDragEnd = () => null,
+	onDragEnter = () => null,
+	onDragOver = () => null,
 	currentExit = false
 }: IUserTeamCard) {
 	const t = useTranslations();
@@ -224,45 +197,6 @@ export function UserTeamCard({
 			{currentExit && (
 				<HorizontalSeparator className="mt-2 !border-primary-light dark:!border-primary-light !border-t-2" />
 			)}
-		</div>
-	);
-}
-
-export function UserTeamCardSkeleton() {
-	return (
-		<div
-			role="status"
-			className="p-4 border divide-y divide-gray-200 shadow rounded-xl animate-pulse dark:divide-gray-700 md:p-6 dark:border-gray-700"
-		>
-			<div className="flex items-center justify-between">
-				<div className="flex items-center space-x-3">
-					<div className="w-5 h-5 mr-8 rounded-[50%] bg-gray-200 dark:bg-gray-700"></div>
-					<div className="w-14 h-14 rounded-[50%] bg-gray-200 dark:bg-gray-700"></div>
-					<div>
-						<div className="w-32 h-3 mb-2 bg-gray-200 rounded-full dark:bg-gray-700"></div>
-					</div>
-				</div>
-				<div className="h-2.5 bg-gray-300 rounded-full dark:bg-gray-600 w-24"></div>
-				<div className="h-2.5 bg-gray-300 rounded-full dark:bg-gray-600 w-24"></div>
-				<div className="h-2.5 bg-gray-300 rounded-full dark:bg-gray-600 w-24"></div>
-				<div className="h-2.5 bg-gray-300 rounded-full dark:bg-gray-700 w-24"></div>
-			</div>
-		</div>
-	);
-}
-
-export function InviteUserTeamSkeleton() {
-	return (
-		<div
-			role="status"
-			className="p-4 mt-3 border divide-y divide-gray-200 shadow rounded-xl animate-pulse dark:divide-gray-700 md:p-6 dark:border-gray-700"
-		>
-			<div className="flex items-center justify-between">
-				<div className="flex items-center space-x-3">
-					<div className="w-5 h-5 mr-8 rounded-[50%] bg-gray-200 dark:bg-gray-700"></div>
-					<div className="w-24 bg-gray-200 h-9 rounded-xl dark:bg-gray-700"></div>
-				</div>
-			</div>
 		</div>
 	);
 }
